custom_requirement.txt
accelerate==1.2.1
addict==2.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==24.3.0
bleach==6.2.0
blinker==1.9.0
blis==1.1.0
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
configargparse==1.7
contexttimer==0.3.3
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash==2.18.2
debugpy==1.8.11
decorator==4.4.2
decord==0.6.0
diffusers==0.14.0
distlib==0.3.9
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.4
fastjsonschema==2.21.1
filelock==3.16.1
flask==3.0.3
fonttools==4.55.3
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.44
h5py==3.12.1
huggingface-hub==0.25.0
identify==2.6.5
idna==3.10
imageio-ffmpeg==0.5.1
imageio==2.36.1
importlib-metadata==8.5.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
joblib==1.4.2
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kaggle==1.6.17
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
lazy-loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
matplotlib==3.10.0
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.11
narwhals==1.21.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open3d==0.19.0
opencv-python-headless==4.5.5.64
opencv-python==4.10.0.84
opendatasets==0.1.22
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==10.4.0
pip==22.0.2
platformdirs==4.3.6
plotly==5.24.1
portalocker==3.1.1
pre-commit==4.0.1
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic-core==2.27.2
pydantic==2.10.4
pydeck==0.9.1
pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
sentencepiece==0.2.0
setuptools==59.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.8.3
srsly==2.5.0
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
thinc==8.3.3
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.27.0
triton==3.1.0
typer==0.15.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.1
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
zipp==3.21.0